+++ Source/UI/NewUI/BillingForm/View/BillingFormCellTextField.swift
@@ -22,9 +22,7 @@ final class BillingFormCellTextField: UITableViewCell {
     }()
 
     private lazy var tapGesture: UITapGestureRecognizer = {
-      let tapGesture = UITapGestureRecognizer(target: self, action: #selector(hideKeyboard(gestureRecognizer:)))
-      tapGesture.cancelsTouchesInView = false
-      return tapGesture
+      UIView.keyboardDismissTapGesture
     }()
 
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
@@ -45,14 +43,6 @@ final class BillingFormCellTextField: UITableViewCell {
     required init?(coder _: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-
-  @objc private func hideKeyboard(gestureRecognizer: UITapGestureRecognizer) {
-    let view = gestureRecognizer.view
-    let loc = gestureRecognizer.location(in: view)
-    let subview = view?.hitTest(loc, with: nil)
-    guard !(subview is SecureDisplayView) else { return }
-    UIApplication.shared.sendAction(#selector(UIApplication.resignFirstResponder), to: nil, from: nil, for: nil)
-  }
 
 }
