+++ Source/UI/NewUI/CommonUI/View/Component/InputView.swift
@@ -46,23 +46,13 @@ class InputView: UIView {
     }()
 
     private lazy var tapGesture: UITapGestureRecognizer = {
-      let tapGesture = UITapGestureRecognizer(target: self, action: #selector(hideKeyboard(gestureRecognizer:)))
-      tapGesture.cancelsTouchesInView = false
-      return tapGesture
+      UIView.keyboardDismissTapGesture
     }()
 
     override init(frame: CGRect) {
       super.init(frame: frame)
       addGestureRecognizer(tapGesture)
       setupViewsInOrder()
-    }
-
-    @objc private func hideKeyboard(gestureRecognizer: UITapGestureRecognizer) {
-      let view = gestureRecognizer.view
-      let loc = gestureRecognizer.location(in: view)
-      let subview = view?.hitTest(loc, with: nil)
-      guard !(subview is SecureDisplayView) else { return }
-      UIApplication.shared.sendAction(#selector(UIApplication.resignFirstResponder), to: nil, from: nil, for: nil)
     }
 
     required init?(coder: NSCoder) {
